script: add auto_bump script to compile on centos7 docker
We didn't have a script to build ready-to-run binaries, which would confuse
some new XiangShan users and developers from the community and even make it
hard to bump them when new features and bug fixes are added to spike or
nemu. So adding this script to get this solved.

Note: Building on CentOS 7 is now a workaround for some EDA environments. I
hope we will use some modern software package manager like Nix someday to
run this software on the EDA environment.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>